Moved the new Scala3-compatible GenIso.fields into the core module as an extension of monocle.Iso.type
